% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_alias.R
\name{mc_alias_set}
\alias{mc_alias_set}
\title{mc alias set}
\usage{
mc_alias_set(
  alias,
  endpoint_url,
  access_key = NA,
  secret_key = NA,
  token = NA,
  storage = c("config", "env"),
  use_aws_env = FALSE
)
}
\arguments{
\item{alias}{a short name for this endpoint}

\item{endpoint_url}{the endpoint url, starting with the scheme, for example
"http://localhost:9000" or "https://play.min.io" or "https://s3.amazonaws.com"}

\item{access_key}{access key (user)}

\item{secret_key}{secret access key}

\item{token}{temporary session token}

\item{storage}{one of "config" (default) or "env", to indicate if the alias setting
should be stored in the config file for the minio client or as a system
environment variable}

\item{use_aws_env}{logical to indicate whether settings for the alias should
be inferred from currently specified AWS environment variables}
}
\value{
Returns the list from \code{\link[processx:run]{processx::run()}}, with components \code{status},
\code{stdout}, \code{stderr}, and \code{timeout}; invisibly.
}
\description{
Set a new alias for the minio client, possibly using env var defaults.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

mc_alias_set()
\dontshow{\}) # examplesIf}
}
\references{
\url{https://docs.min.io/aistor/reference/cli/}.
Note that keys can be omitted for anonymous use.
}
