% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_alias.R
\name{mc_aliases}
\alias{mc_aliases}
\title{Returns all aliases known by the minio client, including those provided
through MC_HOST_* environment variables}
\usage{
mc_aliases(details = FALSE, show_secret = FALSE)
}
\arguments{
\item{details}{logical to indicate whether results should be provided as
a vector with the alias names or as a data frame with full details}

\item{show_secret}{logical to indicate whether to display the secret pass
or to blank it out (the default)}
}
\value{
a vector of aliases or a data frame with alias details
}
\description{
Returns all aliases known by the minio client, including those provided
through MC_HOST_* environment variables
}
