% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_anonymous.R
\name{mc_anonymous_set}
\alias{mc_anonymous_set}
\alias{mc_policy_set}
\title{Set anonymous access policy}
\usage{
mc_anonymous_set(target, policy = c("download", "upload", "public", "private"))
}
\arguments{
\item{target}{Character string specifying the target cloud storage
bucket or object}

\item{policy}{Character string specifying the anonymous access policy.
Must be one of "download", "upload", "public" (upload and download),
or "private".}
}
\value{
None
}
\description{
This function uses the \code{mc} command to set the anonymous access policy for
a specified target.
}
\examples{
\dontrun{
# Set anonymous access policy to download
mc_anonymous_set("alias/bucket/file.txt", policy = "download")

# Set anonymous access policy to upload
mc_anonymous_set("alias/bucket/directory", policy = "upload")

# Set anonymous access policy to public
mc_anonymous_set("alias/bucket/file.txt", policy = "public")

# Set anonymous access policy to private (default policy for new buckets)
mc_anonymous_set("alias/bucket/directory", policy = "private")
}
}
