% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_du.R
\name{mc_du}
\alias{mc_du}
\title{Show disk usage for a target path}
\usage{
mc_du(target, flags = "")
}
\arguments{
\item{target}{alias/bucket to list}

\item{flags}{optional additional flags}
}
\description{
Show disk usage for a target path
}
\details{
for more help, run \code{mc_du("-h")}
}
\examples{
\dontrun{

mc_mb("play/test")
mc_du("play/test")
}
}
