% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_ls.R
\name{mc_ls_tbl}
\alias{mc_ls_tbl}
\title{Directory listing as data frame}
\usage{
mc_ls_tbl(target, recursive = FALSE)
}
\arguments{
\item{target}{Character string specifying the target directory path.}

\item{recursive}{Logical indicating whether to recursively list directories.
Default is \code{FALSE}.}
}
\value{
a data.frame with the directory listing information
}
\description{
This function parses the stdout from \code{mc_ls} and returns the result as a data frame.
}
