% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_mb.R
\name{mc_mb}
\alias{mc_mb}
\title{Create a new S3 bucket using mc command}
\usage{
mc_mb(bucket)
}
\arguments{
\item{bucket}{Character string specifying the name of the bucket to create.}
}
\description{
Create a new S3 bucket using mc command
}
\examples{
\dontrun{
# Create a new bucket named "my-bucket"
mc_mb("my-bucket")
}

}
