% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_rb.R
\name{mc_rb}
\alias{mc_rb}
\title{Remove an S3 bucket using mc command}
\usage{
mc_rb(bucket)
}
\arguments{
\item{bucket}{Character string specifying the name of the bucket to remove}
}
\description{
Remove an S3 bucket using mc command
}
\examples{
\dontrun{
# Create a new bucket named "my-bucket" on the "play" system
mc_mb("play/my-bucket")
mc_rb("play/my-bucket")
}

}
