% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_stat.R
\name{mc_stat}
\alias{mc_stat}
\title{Show object metadata}
\usage{
mc_stat(target, verbose = TRUE, flags = "", details = FALSE)
}
\arguments{
\item{target}{the target specification, for example "play" (an alias) or
"s3/openalex" (a bucket) or "play/test/test.tar.gz" (a specific object) or "R"
(a local path)}

\item{verbose}{logical indicating whether to include some additional information,
by default TRUE}

\item{flags}{string with additional flags to be sent along to minio client,
such as "--recursive"}

\item{details}{logical for returning results as a data frame, by default FALSE}
}
\description{
Object metadata can include information about content type and
various other metadata attributes, if associated with the object.
}
