% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_alias.R
\name{parse_mc_host_env}
\alias{parse_mc_host_env}
\title{Parse a named vector containing MC_HOST_* environment variable(s)}
\usage{
parse_mc_host_env(x, show_secret = TRUE)
}
\arguments{
\item{x}{a named vector - where the name corresponds to the "alias"}

\item{show_secret}{logical indicating whether to blank out or
to include the secret access key (the default)}
}
\value{
a data frame with the same columns as when
}
\description{
In addition to known aliases against S3 servers configured in the
config.json file for the minio client, settings for remote connections
or aliases can also be provided through system environment variables
named "MC_HOST_*" using one string.
}
\details{
This function parses such a string and returns the components as a
data frame.

In minio parlance these settings together constitute an "alias" and the
asterisk in MC_HOST_* represents a placeholder for the corresponding alias.

These MC_HOST_* environment variable string values are similar to database
connection strings; they encode information about protocol/scheme, credentials
and optional session tokens used and server host URL in a single string -
ie settings required for making a connection an S3 server.
}
