% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_alias.R
\name{set_alias_from_aws_env}
\alias{set_alias_from_aws_env}
\title{Creates and activates a MC_HOST_alias environment variable based on AWS
system environment variables.}
\usage{
set_alias_from_aws_env(alias)
}
\arguments{
\item{alias}{string with the alias name for the connection settings}
}
\description{
This function parses current system variable settings for  AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY,
AWS_DEFAULT_REGION, AWS_ENDPOINT_URL and AWS_SESSION_TOKEN and builds a
single MC_HOST_alias environment variable for the same connection and activates
this connection as an alias available to the minio client.
}
