% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{install_mc}
\alias{install_mc}
\title{install the mc client}
\usage{
install_mc(os = "linux", arch = "amd64", path = bin_path())
}
\arguments{
\item{os}{operating system}

\item{arch}{architecture ("amd64" or "ppc64")}

\item{path}{destination where binary is installed.}
}
\description{
install the mc client
}
