% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc}
\alias{mc}
\title{mc}
\usage{
mc(command, ..., path = bin_path())
}
\arguments{
\item{command}{text string of an mc command (starting after the mc ...)}

\item{...}{additional arguments to \verb{[processx::run]}}

\item{path}{location where mc executable will be installed. By default will
use the OS-appropriate storage location.}
}
\value{
pid for the running processx process, invisibly.
}
\description{
The MINIO Client
}
\details{
see \url{https://docs.min.io/docs/minio-client-quickstart-guide.html}
}
