% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc}
\alias{mc}
\title{mc}
\usage{
mc(command, ..., path = bin_path(), verbose = TRUE)
}
\arguments{
\item{command}{space-delimited text string of an mc command
(starting after the mc ...)}

\item{...}{additional arguments to \code{\link[processx:run]{processx::run()}}}

\item{path}{location where mc executable will be installed. By default will
use the OS-appropriate storage location.}

\item{verbose}{print output?}
}
\value{
Returns the list from \code{\link[processx:run]{processx::run()}}, with components \code{status},
\code{stdout}, \code{stderr}, and \code{timeout}; invisibly.
}
\description{
The MINIO Client
}
\details{
This function forms the basis for all other available commands.
This utility can run any \code{mc} command supported by the official minio client,
see \url{https://min.io/docs/minio/linux/reference/minio-mc.html}.
The R package provides wrappers only for the most common use cases,
which provide a more natural R syntax and native documentation.
}
