% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_alias.R
\name{mc_alias_ls}
\alias{mc_alias_ls}
\title{List all configured aliases}
\usage{
mc_alias_ls(alias = "")
}
\arguments{
\item{alias}{optional argument, display only specified alias}
}
\value{
Configured aliases, including secret keys!
}
\description{
List all configured aliases
}
\details{
Note that all available
}
\seealso{
mc
}
