% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc_alias_set}
\alias{mc_alias_set}
\title{mc alias set}
\usage{
mc_alias_set(
  alias = "minio",
  access_key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret_key = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  scheme = "https",
  endpoint = Sys.getenv("AWS_S3_ENDPOINT", "s3.amazonaws.com")
)
}
\arguments{
\item{alias}{a short name for this endpoint, default is \code{minio}}

\item{access_key}{accesss key (user), will be read from AWS env vars by default}

\item{secret_key}{secret access key, will be read from AWS env vars by default}

\item{scheme}{https or http (e.g. for local machine only)}

\item{endpoint}{the endpoint domain name}
}
\description{
Set a new alias for the minio client, possibly using env var defaults.
}
\details{
see \url{https://docs.min.io/docs/minio-client-quickstart-guide.html#add-a-cloud-storage-service}
}
