% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_ls.R
\name{mc_ls}
\alias{mc_ls}
\title{List files and directories using mc command}
\usage{
mc_ls(target, recursive = FALSE, show_fullpath = TRUE)
}
\arguments{
\item{target}{character vector specifying the target directory path(s).}

\item{recursive}{Logical indicating whether to recursively list directories.
Default is \code{FALSE}.}

\item{show_fullpath}{logical, by default TRUE, a column with the
full path is included in the listing}
}
\value{
a data.frame with the directory listing information
}
\description{
This function uses the \code{mc} command to list files and directories
at the specified target location and returns a data frame with the results.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# list all buckets on play server
mc_ls("play/")

\dontshow{\}) # examplesIf}
}
