% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_ls.R
\name{mc_ls}
\alias{mc_ls}
\title{List files and directories using mc command}
\usage{
mc_ls(target, recursive = FALSE, flags = "")
}
\arguments{
\item{target}{Character string specifying the target directory path.}

\item{recursive}{Logical indicating whether to recursively list directories.
Default is \code{FALSE}.}

\item{flags}{Additional flags to be passed to the \code{ls} command.
Default is an empty string.}
}
\value{
list of files
}
\description{
This function uses the \code{mc} command to list files and directories
at the specified target location.
}
\examples{
\dontrun{
# List files and directories in a directory
mc_ls("path/to/directory")

# List files and directories recursively in a directory
mc_ls("path/to/directory", recursive = TRUE)

# List files and directories with additional flags
mc_ls("path/to/directory", flags = "-l")
}

}
